// frontend/prism/app/services/api/models/public-user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Prism API
 * API for Prism project
 *
 * The version of the OpenAPI document: v1
 * Contact: mercysquadrant@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PublicUser
 */
export interface PublicUser {
    /**
     * 
     * @type {number}
     * @memberof PublicUser
     */
    'id'?: number;
    /**
     * Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.
     * @type {string}
     * @memberof PublicUser
     */
    'username': string;
    /**
     * 
     * @type {boolean}
     * @memberof PublicUser
     */
    'is_artist'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof PublicUser
     */
    'profile_picture'?: string | null;
}